+++ CHANGELOG.md
@@ -15,8 +15,8 @@ Changelog follow https://keepachangelog.com/ format.
         *   Breaking: Remove `widget=True` argument to  (widget always enabled).
         *   Add `expanded: bool` kwargs to control whether the widget is
             expanded or collapsed by default.
-        *   Breaking: Remove `keep_proto` kwargs from `clear_cached_modules` and
-            `import_proto` kwargs from `adhoc` (proto always supported)
+    *   Breaking: Remove `keep_proto` kwargs from `clear_cached_modules` and
+        `import_proto` kwargs from `adhoc` (proto always supported)
 *   `enp`:
     *   Make array spec (e.g. `etree.spec_like()`) hashable.
 *   `epath`: